pyproject.toml
[project]
name = "ai-hedge-fund"
version = "1.0.0"
description = "An AI-powered hedge fund that uses multiple agents to make trading decisions"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain>=0.3.7,<0.4.0",
    "langchain-anthropic==0.3.5",
    "langchain-groq==0.2.3",
    "langchain-openai>=0.3.5,<0.4.0",
    "langchain-deepseek>=0.1.2,<0.2.0",
    "langchain-ollama==0.3.6",
    "langgraph>=1.0.0,<2.0.0",
    "pandas>=2.1.0,<3.0.0",
    "numpy>=1.26.4",
    "python-dotenv==1.0.0",
    "matplotlib>=3.9.2,<4.0.0",
    "tabulate>=0.9.0,<0.10.0",
    "colorama>=0.4.6,<0.5.0",
    "questionary>=2.1.0,<3.0.0",
    "rich>=13.9.4,<14.0.0",
    "langchain-google-genai>=2.0.11,<3.0.0",
    "fastapi[standard]>=0.104.0,<0.105.0",
    "fastapi-cli>=0.0.7,<0.1.0",
    "pydantic>=2.4.2,<3.0.0",
    "httpx>=0.27.0,<0.28.0",
    "sqlalchemy>=2.0.22,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
    "langchain-gigachat>=0.3.12,<0.4.0",
    "langchain-xai>=0.2.5,<0.3.0",
    "sniffio>=1.3.1",
    "uvicorn>=0.40.0",
    "ruff>=0.0.288",
    "langfuse>=3.0.0",
    "langsmith>=0.2.0",
    "langchain-core>=0.3.83",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "ruff>=0.0.288,<0.0.289"
]

[project.scripts]
backtester = "src.backtesting.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "app*"]

[tool.black]
line-length = 128
target-version = ['py313']
